Aufenthalt und Interfaces geändert

+++ Project/src/projekt/fhv/teama/classes/zimmer/IZimmerpreis.java
@@ -7,6 +7,6 @@ public interface IZimmerpreis extends IDao {
 	public IKategorie getKategorie();
 	public void setSaison(ISaison saison);
 	public ISaison getSaison();
-	public void setPreis(double preis);
-	public double getPreis();
+	public void setPreis(float preis);
+	public float getPreis();
 }
